OCPBUGS-48688: propogate MANAGED_SERVICE env variable to HCCO v2
