+++ control-plane-operator/controllers/hostedcontrolplane/v2/configoperator/deployment.go
@@ -12,74 +12,7 @@ import (
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 )
-
-const (
-	ComponentName = "hosted-cluster-config-operator"
-)
-
-var _ component.ComponentOptions = &HCCO{}
-
-type HCCO struct {
-	registryOverrides               map[string]string
-	openShiftImageRegistryOverrides map[string][]string
-}
-
-// IsRequestServing implements controlplanecomponent.ComponentOptions.
-func (h *HCCO) IsRequestServing() bool {
-	return false
-}
-
-// MultiZoneSpread implements controlplanecomponent.ComponentOptions.
-func (h *HCCO) MultiZoneSpread() bool {
-	return false
-}
-
-// NeedsManagementKASAccess implements controlplanecomponent.ComponentOptions.
-func (h *HCCO) NeedsManagementKASAccess() bool {
-	return true
-}
-
-func NewComponent(registryOverrides map[string]string, openShiftImageRegistryOverrides map[string][]string) component.ControlPlaneComponent {
-	hcco := &HCCO{
-		registryOverrides:               registryOverrides,
-		openShiftImageRegistryOverrides: openShiftImageRegistryOverrides,
-	}
-
-	return component.NewDeploymentComponent(ComponentName, hcco).
-		WithAdaptFunction(hcco.AdaptDeployment).
-		WithManifestAdapter(
-			"podmonitor.yaml",
-			component.WithAdaptFunction(adaptPodMonitor),
-		).
-		WithManifestAdapter(
-			"role.yaml",
-			component.WithAdaptFunction(adaptRole),
-		).
-		InjectAvailabilityProberContainer(util.AvailabilityProberOpts{
-			KubeconfigVolumeName: "kubeconfig",
-			RequiredAPIs: []schema.GroupVersionKind{
-				{Group: "imageregistry.operator.openshift.io", Version: "v1", Kind: "Config"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Infrastructure"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "DNS"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Ingress"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Network"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Proxy"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Build"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Image"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "Project"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "ClusterVersion"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "FeatureGate"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "ClusterOperator"},
-				{Group: "config.openshift.io", Version: "v1", Kind: "OperatorHub"},
-				{Group: "operator.openshift.io", Version: "v1", Kind: "Network"},
-				{Group: "operator.openshift.io", Version: "v1", Kind: "CloudCredential"},
-				{Group: "operator.openshift.io", Version: "v1", Kind: "IngressController"},
-			},
-		}).
-		Build()
-}
 
 const (
 	kubeconfigVolumeName      = "kubeconfig"
@@ -87,7 +20,7 @@ const (
 	clusterSignerCAVolumeName = "cluster-signer-ca"
 )
 
-func (h *HCCO) AdaptDeployment(cpContext component.WorkloadContext, deployment *appsv1.Deployment) error {
+func (h *hcco) adaptDeployment(cpContext component.WorkloadContext, deployment *appsv1.Deployment) error {
 	versions, err := cpContext.ReleaseImageProvider.ComponentVersions()
 	if err != nil {
 		return fmt.Errorf("failed to get component versions: %w", err)
@@ -139,6 +72,13 @@ func (h *HCCO) AdaptDeployment(cpContext component.WorkloadContext, deployment *
 					Value: "1",
 				},
 			)
+		}
+		if len(os.Getenv("MANAGED_SERVICE")) > 0 {
+			c.Env = append(c.Env,
+				corev1.EnvVar{
+					Name:  "MANAGED_SERVICE",
+					Value: os.Getenv("MANAGED_SERVICE"),
+				})
 		}
 	})
